Add documentation, rename parameters, and fix accessibility levels.
