+++ src/FluentFiles.Delimited.Attributes/Properties/AssemblyInfo.cs
@@ -1,4 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("FluentFiles.FixedLength.Attributes")]
 [assembly: InternalsVisibleTo("FluentFiles.Tests")]
