+++ src/FluentFiles.FixedLength.Attributes/Properties/AssemblyInfo.cs
@@ -1,4 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("FluentFiles.Delimited.Attributes")]
 [assembly: InternalsVisibleTo("FluentFiles.Tests")]
