+++ src/FluentFiles.FixedLength/IFixedLayout.cs
@@ -2,6 +2,10 @@
 {
     using FluentFiles.Core;
 
+    /// <summary>
+    /// Describes the mapping from a fixed-length file record to a target type.
+    /// </summary>
+    /// <typeparam name="TTarget">The type that a record maps to.</typeparam>
     public interface IFixedLayout<TTarget> :
         IFixedLengthLayoutDescriptor,
         ILayout<TTarget, IFixedFieldSettingsContainer, IFixedFieldSettingsBuilder, IFixedLayout<TTarget>>
